Slim down function and dependencies
We're hitting our Lambda function and layer limit, so we need to trim
the archives we're uploading.

The `media/` directory is no longer needed at all now we're using the
DC Base Theme.

We can use a newer version of serverless-python-requirements that will
allow us to exclude certain directories from the requirements archive.
We don't need the translations from `localflavor`, just a GB Postcode
field.