Fix leaflet JSON API
We've removed the `image_text` field which was unpopulated, so we
shouldn't try and serialise it.

Leaflets don't have an `owner` so don' t use it to check permissions.

Use a relative URL for the JSON link on the leaflet page, the
`HTTP_HOST` on Lambda will not be the real EL URL and won't work.