Use debug mode for live_server
This is needed because we don't collect static when playwright interacts with the live_server fixture.

Without calling collect static we can't serve some static assets, so we can't test that the front end is working.

The downside of this is that the live server might act differently when in debug mode. There isn't a lot we can do about this,
and we should rely on deplyed integration testing for things like 404s when DEBUG=False.