Move party model to YNR Party fields
This is step one in removing the party model.

We currently have two ways of storing the party of a leaflet:

1. The (old) Party model that requires us syncing parties with a 3rd party data source
2. Pulling in the Party data from YNR when we look up a ballot as part of the uploader.

New leaflets use method two, so this migration moves old data inline with new data.

It's very possible we decide to denormalize this again in future, but for now I want to reduce complexity and move all data to the same system.

I've verified with a prod export that there is no missing data using this method.