Remove Constituency model
We're not writing to this model any more, and even if we were it's not
clear that it's useful: it's the 2010 boundaries with no ability to
version them. We also might not want to filter by constituency anyway.

As we have the postcode for every leaflet, we can re-make this in a way
that we think is useful, at some point later.